+++ src/iofwd/frontend/IOFWDReadRequest.hh
@@ -38,7 +38,7 @@ private:
    zoidfs::zoidfs_handle_t handle_;
    uint32_t mem_count_;
    char ** mem_starts_;
-   uint32_t * mem_sizes_;
+   uint64_t * mem_sizes_;
    uint32_t file_count_;
    uint64_t * file_starts_;
    uint64_t * file_sizes_;
